fix undefined index

+++ zp-core/admin_options/theme.php
@@ -80,7 +80,7 @@ function saveOptions() {
 	if ($themeswitch) {
 		$notify = '?switched';
 	} else {
-		if ($_POST['savethemeoptions'] == 'reset') {
+		if (isset($_POST['savethemeoptions']) && $_POST['savethemeoptions'] == 'reset') {
 			$sql = 'DELETE FROM ' . prefix('options') . ' WHERE `theme`=' . db_quote($themename);
 			if ($themealbum) {
 				$sql .= ' AND `ownerid`=' . $themealbum->getID();
@@ -236,7 +236,7 @@ function getOptionContent() {
 								<p class="buttons">
 									<button type="submit" value="<?php echo gettext('Apply') ?>"><img src="images/pass.png" alt="" /><strong><?php echo gettext("Apply"); ?></strong></button>
 									<button type="button" value="<?php echo gettext('Revert to default') ?>" onclick="$('#savethemeoptions').val('reset');
-											$('#themeoptionsform').submit();"><img src="images/refresh.png" alt="" /><strong><?php echo gettext("Revert to default"); ?></strong></button>
+														$('#themeoptionsform').submit();"><img src="images/refresh.png" alt="" /><strong><?php echo gettext("Revert to default"); ?></strong></button>
 									<button type="reset" value="<?php echo gettext('reset') ?>"><img src="images/reset.png" alt="" /><strong><?php echo gettext("Reset"); ?></strong></button>
 								</p>
 							</td>
@@ -523,7 +523,7 @@ function getOptionContent() {
 								<p class="buttons">
 									<button type="submit" value="<?php echo gettext('Apply') ?>"><img src="images/pass.png" alt="" /><strong><?php echo gettext("Apply"); ?></strong></button>
 									<button type="button" value="<?php echo gettext('Revert to default') ?>" onclick="$('#savethemeoptions').val('reset');
-											$('#themeoptionsform').submit();"><img src="images/refresh.png" alt="" /><strong><?php echo gettext("Revert to default"); ?></strong></button>
+														$('#themeoptionsform').submit();"><img src="images/refresh.png" alt="" /><strong><?php echo gettext("Revert to default"); ?></strong></button>
 									<button type="reset" value="<?php echo gettext('reset') ?>"><img src="images/reset.png" alt="" /><strong><?php echo gettext("Reset"); ?></strong></button>
 								</p>
 							</td>
